requirements_tfidf.txt

streamlit==1.36.0
umap-learn==0.5.6
hdbscan==0.8.38
scikit-learn==1.4.2
plotly==5.22.0
pandas==2.2.2
numpy==1.26.4
